[rename] 더미데이터 파일명 변경 dummy -> Friends
